Import of BOBTFISH/Test-EOL-1.4 from CPAN.
gitpan-cpan-distribution: Test-EOL
gitpan-cpan-version:      1.4
gitpan-cpan-path:         BOBTFISH/Test-EOL-1.4.tar.gz
gitpan-cpan-author:       BOBTFISH
gitpan-cpan-maturity:     released

+++ t/12-fail.t
@@ -13,59 +13,58 @@ $inc = "-I $inc" if $inc;
     open my $fh, '<', $filename or die $!;
     binmode( $fh, ':raw' );
     my $content = <$fh>;
-    is( $content, ascii_string(), 'Data written to file is there when we look for it later' ); 
+    is( $content, ascii_string(), 'Data written to file is there when we look for it later' );
 
 }
 {
     my $dir = make_bad_file_1();
-    my (undef, $outfile) = tempfile();
-    ok( `$perl $inc -MTest::EOL -e "all_perl_files_ok( '$dir' )" 2>&1 > $outfile` );
-    local $/ = undef;
-    open my $fh, '<', $outfile or die $!;
-    my $content = <$fh>;
-    like( $content, qr/^not ok 1 - No incorrect line endings in '[^']*' \Qon line 4: [\r]/m, 'windows EOL found in tmp file 1' );
-    unlink $outfile;
+    run_ok( "all_perl_files_ok( '$dir' )",
+            qr/^not ok 1 - No incorrect line endings in '[^']*' \Qon line 5: [\r]/m,
+            'windows EOL found in tmp file 1' );
 }
 {
     my $dir = make_bad_file_2();
-    my (undef, $outfile) = tempfile();
-    ok( `$perl $inc -MTest::EOL -e "all_perl_files_ok( '$dir' )" 2>&1 > $outfile` );
-    open my $fh, '<', $outfile or die $!;
-    local $/ = undef;
-    my $content = <$fh>;
-    like( $content, qr/^not ok 1 - No incorrect line endings in '[^']*' \Qon line 8: [\r][\r][\r][\r][\r][\r][\r]/m, 'windows EOL found in tmp file2 ' );
-    unlink $outfile;
+    run_ok( "all_perl_files_ok( '$dir' )",
+            qr/^not ok 1 - No incorrect line endings in '[^']*' \Qon line 8: [\r][\r][\r][\r][\r][\r][\r]/m,
+            'windows EOL found in tmp file2 ' );
 }
 {
     my ($dir, $file) = make_bad_file_3();
-    my (undef, $outfile) = tempfile();
-    ok( `$perl $inc -MTest::EOL -e "all_perl_files_ok( '$file' )" 2>&1 > $outfile` );
-    open my $fh, '<', $outfile or die $!;
-    local $/ = undef;
-    my $content = <$fh>;
-    like( $content, qr/^not ok 1 - No incorrect line endings in '[^']*' \Qon line 9: [\r][\r][\r]/m, 'windows EOL found in tmp file 3' );
-    unlink $outfile;
+    run_ok( "all_perl_files_ok( '$file' )",
+            qr/^not ok 1 - No incorrect line endings in '[^']*' \Qon line 1: [\r] /m,
+            'windows EOL found in tmp file 3' );
 }
 
 {
     my $dir = make_bad_file_4();
+    run_ok( "all_perl_files_ok({trailing_whitespace => 1}, '$dir' )",
+            # Note that line number will be 13
+            qr/^not ok 1 - No incorrect line endings in '[^']*' \Qon line 12: [\s][\t][\s][\s]/m,
+            'Trailing ws EOL found in tmp file 4' );
+}
+
+sub run_ok {
+    my ($code, $match, $test_name) = @_;
+    my $line = (caller)[2];
+    die "code containing double quotes is not portable on Win32 in one-liners" if $code =~ /"/;
     my (undef, $outfile) = tempfile();
-    ok( `$perl $inc -MTest::EOL -e "all_perl_files_ok({trailing_whitespace => 1}, '$dir' )" 2>&1 > $outfile` );
-    open my $fh, '<', $outfile or die $!;
+    is( `$perl $inc -MTest::EOL -e "$code" > $outfile 2>&1`, '', "test sub program at line $line: output redirected" );
+    is( $? >> 8, 1, "test sub program at line $line: exit code is 1" );
     local $/ = undef;
+    open my $fh, '<', $outfile or die $!;
     my $content = <$fh>;
-    like( $content, qr/^not ok 1 - No incorrect line endings in '[^']*' \Qon line 13: [\s][\t][\s][\s]/m, 'Trailing ws EOL found in tmp file 4' );
+    like( $content, $match, $test_name );
     unlink $outfile;
 }
 
-sub ascii_string { 
+sub ascii_string {
   my $o = "<before \r\n between \r\n after \n normal >";
   return $o x 3;
 }
 
-sub make_raw_badfile { 
-  my $tmpdir = tempdir( CLEANUP => 1 ); 
-  my ( $fh, $filename ) = tempfile( DIR => $tmpdir, SUFFIX =>  '.tXt' ); 
+sub make_raw_badfile {
+  my $tmpdir = tempdir( CLEANUP => 1 );
+  my ( $fh, $filename ) = tempfile( DIR => $tmpdir, SUFFIX =>  '.tXt' );
   binmode $fh, ':raw';
   print $fh ascii_string();
   close $fh;
@@ -81,6 +80,7 @@ sub make_bad_file_1 {
 #!perl
 
 sub main {
+    # Note that the generated file will have the CRLF expanded in the source
     print "Hello!\r\n";
 }
 DUMMY
@@ -101,11 +101,11 @@ sub make_bad_file_2 {
 =head1 NAME
 
 test.pL -	A test script
-
+\r\r\r\r\r\r\r\r
 =cut
 
 sub main {
-    print "Hello!\n";
+    print "Hello!\\n";
 }
 DUMMY
   return $tmpdir;
@@ -116,18 +116,18 @@ sub make_bad_file_3 {
   my ($fh, $filename) = tempfile( DIR => $tmpdir, SUFFIX => '.pm' );
   binmode $fh, ':raw';
   print $fh <<"DUMMY";
-use strict;
-
-package My::Test;
-
-sub new {
-    my (\$class) = @_;
-    my \$self = bless { }, \$class;
-    return \$self;
-}
-
-
-1;
+use strict;\r
+\r
+package My::Test;\r
+\r
+sub new {\r
+    my (\$class) = \@_;\r
+    my \$self = bless { }, \$class;\r
+    return \$self;\r
+}\r
+\r
+\r
+1;\r
 DUMMY
   close $fh;
   return ($tmpdir, $filename);
@@ -137,7 +137,7 @@ sub make_bad_file_4 {
   my $tmpdir = tempdir( CLEANUP => 1 );
   my ($fh, $filename) = tempfile( DIR => $tmpdir, SUFFIX => '.pL' );
   binmode $fh, ':raw';
-  print $fh <<"DUMMY";
+  print $fh <<'DUMMY';
 #!perl
 
 =pod
@@ -151,7 +151,7 @@ test.pL -	A test script
 sub main {
 DUMMY
 
-print $fh qq{    print "Hello!\n"; \t  \n}; # <-- whitespace
+print $fh qq{    print "Hello!\\n"; \t  \n}; # <-- whitespace
 print $fh '}';
 
   return $tmpdir;
